+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to `ar-php-laravel` will be documented in this file
 
-## 1.0.0 - 2021-04-06
+## 1.3.0 - 2021-05-08
+- remove `trait SearchAr` in favoer of using macrables
+- adding macros for Query Builder and Eloquent Builder
 
+## 1.0.0 - 2021-04-06
 - initial release